+++ sync-for-payables/src/codatsyncpayables/models/shared/pushoptionproperty.py
@@ -2,9 +2,9 @@
 
 from __future__ import annotations
 import dataclasses
-from ..shared import pushoptionchoice as shared_pushoptionchoice
-from ..shared import pushoptiontype as shared_pushoptiontype
-from ..shared import pushvalidationinfo as shared_pushvalidationinfo
+from .pushoptionchoice import PushOptionChoice
+from .pushoptiontype import PushOptionType
+from .pushvalidationinfo import PushValidationInfo
 from codatsyncpayables import utils
 from dataclasses_json import Undefined, dataclass_json
 from typing import Dict, List, Optional
@@ -19,10 +19,10 @@ class PushOptionProperty:
     r"""The property's display name."""
     required: bool = dataclasses.field(metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('required') }})
     r"""The property is required if `True`."""
-    type: shared_pushoptiontype.PushOptionType = dataclasses.field(metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('type') }})
+    type: PushOptionType = dataclasses.field(metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('type') }})
     r"""The option type."""
-    options: Optional[List[shared_pushoptionchoice.PushOptionChoice]] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('options') }})
+    options: Optional[List[PushOptionChoice]] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('options') }})
     properties: Optional[Dict[str, PushOptionProperty]] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('properties') }})
-    validation: Optional[shared_pushvalidationinfo.PushValidationInfo] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('validation'), 'exclude': lambda f: f is None }})
+    validation: Optional[PushValidationInfo] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('validation'), 'exclude': lambda f: f is None }})