+++ sync-for-payables/src/codatsyncpayables/models/shared/withholdingtax_items.py
@@ -9,7 +9,7 @@ from decimal import Decimal
 
 @dataclass_json(undefined=Undefined.EXCLUDE)
 @dataclasses.dataclass
-class WithholdingTaxitems:
+class WithholdingTaxItems:
     amount: Decimal = dataclasses.field(metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('amount'), 'encoder': utils.decimalencoder(False, False), 'decoder': utils.decimaldecoder }})
     r"""Amount of tax withheld."""
     name: str = dataclasses.field(metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('name') }})